Extract json/serdes i64 lowering into host_json_serdes.rs (#1254)
Follows the #1383 pure-function move pattern: child module starts with
'use super::*;', parent adds 'mod host_json_serdes;' and
'pub(crate) use host_json_serdes::*;' next to the existing host_* wiring.

Moved the focused json serdes lowering slice (definitions only; call sites
stay in the parent and resolve through the re-export):
- predicates: is_i64_std_json_wrapper, is_i64_json_parse_*_name,
  is_i64_json_stringify_*_name
- helpers: i64_json_safe_string_len_key, lower_i64_json_escaped_string_len_expr,
  lower_i64_json_safe_string_len_expr
- lowerers: lower_i64_json_stringify_string_line_stmts,
  lower_i64_json_stringify_string_line_stmts_with_written

I64StaticBindings fields/population, json_serdes_* runtime helpers, panic/
logging (lower_i64_log_json_*) and the recursive expr/stmt hub stay in the
parent, reachable from the child via 'use super::*;'.

docs/compiler-source-decomposition-plan.md ratchet update (#1382): add
host_json_serdes.rs ceiling, lower cranelift_backend.rs / summary.top_file_lines
/ summary.top_file_line_share ceilings, and mark the json/serdes slice.

Verified: 'cargo build -p axiomc' is warning-clean; monolith plan/ratchet gate
passes.
